requirements.txt
streamlit>=1.24.0
transformers>=4.30.0
torch>=2.0.0
flask>=2.0.0
flask-cors>=3.0.10
python-dotenv>=0.19.0
requests>=2.26.0
pandas>=1.3.0
numpy>=1.21.0
beautifulsoup4>=4.9.3
ccxt>=2.0.0
plotly>=5.15.0
yfinance>=0.2.0